Move JavaScript files from /js directory to /wp-content/themes/likealunatic30 directory
