+++ .gitignore
@@ -1,2 +1,3 @@
 tmp
 _design
+old_wordpress
